requirements.txt
fastapi
uvicorn
python-multipart
whisper-cpp-python
llama-cpp-python
pyttsx3
